adding hardware code
p02.ino is loaded onto the feather and p02.py is run on the PC. For the PC code, you will need the Adafruit_BluefruitLE code, included in this commit. For setting up the feather with the code, you will need to set up the libraries listed at the top of the p02.ino file. I've included most of them in this commit.